components/toxic-multilabel-trainer/requirements.txt
scipy
scikit-learn
joblib
google-cloud-storage
pandas
requests